+++ NEWS.md
@@ -1,3 +1,14 @@
+# fmesher 0.1.1
+
+* Simplify LICENSE information in the R package DESCRIPTION
+
+* Fix example for legacy `plot_PolySet` method
+
+* Fix C++ unused-variable warning detected by clang-tidy version 16.0.6
+
+* Add workaround for `sf::st_buffer` not supporting negative distance
+  for `s2` with longlat coordinates. Fixes #5
+
 # fmesher 0.1.0
 
 * Full conversion of the fmesher interface from the standalone binary in R-INLA